Refactor out query and cursor processing
